Completely refactor the calendar scraper to api calling approach
