+++ scrapers/calendar.go
@@ -5,362 +5,197 @@
 package scrapers
 
 import (
-	"context"
+	"bytes"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"log"
+	"net/http"
 	"os"
-	"regexp"
 	"time"
 
 	"github.com/UTDNebula/api-tools/utils"
 	"github.com/UTDNebula/nebula-api/api/schema"
-	"github.com/chromedp/cdproto/cdp"
-	"github.com/chromedp/cdproto/runtime"
-	"github.com/chromedp/chromedp"
 	"go.mongodb.org/mongo-driver/bson/primitive"
 )
 
-const CALENDAR_LINK string = "https://calendar.utdallas.edu/calendar"
+// Structure of the API response
+type RawEvent struct {
+	Event map[string]interface{} `json:"event"`
+}
 
-var trailingSpaceRegex *regexp.Regexp = regexp.MustCompile(`(\s{2,}?\s{2,})|(\n)`)
-var leadingSpaceRegex *regexp.Regexp = regexp.MustCompile(`^\s+`)
+type APICalendarResponse struct {
+	Events []RawEvent        `json:"events"`
+	Page   map[string]int    `json:"page"`
+	Date   map[string]string `json:"date"`
+}
 
+// Get the calendar data through API instead of scraping from website
 func ScrapeCalendar(outDir string) {
-
-	chromedpCtx, cancel := utils.InitChromeDp()
-	defer cancel()
-
 	err := os.MkdirAll(outDir, 0777)
 	if err != nil {
 		panic(err)
 	}
+	cli := http.Client{Timeout: 15 * time.Second}
+	var calendarData APICalendarResponse
 
-	events := []schema.Event{}
+	// Get the total number of pages
+	log.Printf("Getting the number of pages...")
+	if err := scrapeAndUnmarshal(&cli, 0, &calendarData); err != nil {
+		panic(err)
+	}
+	numPages := calendarData.Page["total"]
+	log.Printf("The number of pages is %d!\n\n", numPages)
 
-	log.Printf("Scraping event page links")
-	// Grab all links to event pages
-	var pageLinks []string = []string{}
-	_, err = chromedp.RunResponse(chromedpCtx,
-		chromedp.Navigate(CALENDAR_LINK),
-		chromedp.QueryAfter(".em-card_image > a",
-			func(ctx context.Context, _ runtime.ExecutionContextID, nodes ...*cdp.Node) error {
-				for _, node := range nodes {
-					href, hasHref := node.Attribute("href")
-					if !hasHref {
-						return errors.New("event card was missing an href")
-					}
-					pageLinks = append(pageLinks, href)
-				}
-				return nil
-			},
-		),
-	)
+	var events []schema.Event
+	for page := range numPages {
+		log.Printf("Scraping events of page %d...", page+1)
+		if err := scrapeAndUnmarshal(&cli, page+1, &calendarData); err != nil {
+			panic(err)
+		}
+		log.Printf("Scraped events of page %d successfully!\n", page+1)
+
+		log.Printf("Parsing the events of page %d...", page+1)
+		for _, rawEvent := range calendarData.Events {
+			// Parse the time
+			eventInstance := toMap(toMap(toSlice(rawEvent.Event["event_instances"])[0])["event_instance"])
+			startTime := parseTime(toString(eventInstance["start"]))
+			endTime := startTime
+			if toString(eventInstance["end"]) != "" {
+				endTime = parseTime(toString(eventInstance["end"]))
+			}
+
+			location := utils.TrimWhitespace(fmt.Sprintf("%s, %s", toString(rawEvent.Event["location_name"]), toString(rawEvent.Event["room_number"])))
+
+			// Parse the event types, event topic, and event target audience
+			filters := toMap(rawEvent.Event["filters"])
+			eventTypes := []string{}
+			eventTopics := []string{}
+			targetAudiences := []string{}
+
+			rawTypes := toSlice(filters["event_types"])
+			for _, rawType := range rawTypes {
+				eventTypes = append(eventTypes, toString(toMap(rawType)["name"]))
+			}
+
+			rawAudiences := toSlice(filters["event_target_audience"])
+			for _, audience := range rawAudiences {
+				targetAudiences = append(targetAudiences, toString(toMap(audience)["name"]))
+			}
+
+			rawTopics := toSlice(filters["event_topic"])
+			for _, topic := range rawTopics {
+				eventTopics = append(eventTopics, toString(toMap(topic)["name"]))
+			}
+
+			// Parse the event departments, and tags
+			departments := []string{}
+			tags := []string{}
+
+			rawTags := toSlice(rawEvent.Event["tags"])
+			for _, tag := range rawTags {
+				tags = append(tags, tag.(string))
+			}
+
+			rawDeparments := toSlice(rawEvent.Event["departments"])
+			for _, deparment := range rawDeparments {
+				departments = append(departments, toMap(deparment)["name"].(string))
+			}
+
+			// Parse the contact info, =ote that some events won't have contact phone number
+			rawContactInfo := toMap(rawEvent.Event["custom_fields"])
+			contactInfo := [3]string{}
+			for i, infoField := range []string{
+				"contact_information_name", "contact_information_email", "contact_information_phone",
+			} {
+				contactInfo[i] = toString(rawContactInfo[infoField])
+			}
+
+			events = append(events, schema.Event{
+				Id:                 primitive.NewObjectID(),
+				Summary:            toString(rawEvent.Event["title"]),
+				Location:           location,
+				StartTime:          startTime,
+				EndTime:            endTime,
+				Description:        toString(rawEvent.Event["description_text"]),
+				EventType:          eventTypes,
+				TargetAudience:     targetAudiences,
+				Topic:              eventTopics,
+				EventTags:          tags,
+				EventWebsite:       toString(rawEvent.Event["url"]),
+				Department:         departments,
+				ContactName:        contactInfo[0],
+				ContactEmail:       contactInfo[1],
+				ContactPhoneNumber: contactInfo[2],
+			})
+		}
+		log.Printf("Parsed the events of page %d successfully!\n\n", page+1)
+	}
+
+	if err := utils.WriteJSON(fmt.Sprintf("%s/events.json", outDir), events); err != nil {
+		panic(err)
+	}
+	log.Printf("Finished parsing %d events successfully!\n\n", len(events))
+}
+
+// Scrape the data from the api and unmarshal it to response data
+func scrapeAndUnmarshal(client *http.Client, page int, data *APICalendarResponse) error {
+	// Call API to get the byte data
+	calendarUrl := fmt.Sprintf("https://calendar.utdallas.edu/api/2/events?days=365&pp=100&page=%d", page)
+	req, err := http.NewRequest("GET", calendarUrl, nil)
+	if err != nil {
+		return err
+	}
+	res, err := client.Do(req)
+	if err != nil {
+		return err
+	}
+	if res != nil && res.StatusCode != 200 {
+		return fmt.Errorf("ERROR: Non-200 status is returned, %s", res.Status)
+	}
+
+	// Unmarshal bytes to the response data
+	buffer := bytes.Buffer{}
+	if _, err = buffer.ReadFrom(res.Body); err != nil {
+		return err
+	}
+	res.Body.Close()
+	if err = json.Unmarshal(buffer.Bytes(), &data); err != nil {
+		return err
+	}
+	return nil
+}
+
+// Casting an interface{} to an slice of interface{}
+func toSlice(data interface{}) []interface{} {
+	if array, ok := data.([]interface{}); ok {
+		return array
+	}
+	return nil
+}
+
+// Casting an interface{} to map from string to interface{}
+func toMap(data interface{}) map[string]interface{} {
+	if dataMap, ok := data.(map[string]interface{}); ok {
+		return dataMap
+	}
+	return nil
+}
+
+// Casting an interface{} to string, if the data is nil then string is ""
+func toString(data interface{}) string {
+	if data != nil {
+		if dataString, ok := data.(string); ok {
+			return dataString
+		}
+	}
+	return ""
+}
+
+// Parse string time
+func parseTime(stringTime string) time.Time {
+	parsedTime, err := time.Parse(time.RFC3339, stringTime)
 	if err != nil {
 		panic(err)
 	}
-	log.Printf("Scraped event page links!")
-	for _, page := range pageLinks {
-		// Print the links of the page to check
-		log.Println(page)
-	}
-
-	for _, page := range pageLinks {
-		// Navigate to page and get page summary
-		summary := ""
-		_, err := chromedp.RunResponse(chromedpCtx,
-			chromedp.Navigate(page),
-			chromedp.QueryAfter(".em-card_title",
-				func(ctx context.Context, _ runtime.ExecutionContextID, nodes ...*cdp.Node) error {
-					if len(nodes) != 0 {
-						summary = trailingSpaceRegex.ReplaceAllString(getNodeText(nodes[0]), "")
-					}
-					return nil
-				}, chromedp.AtLeast(0),
-			),
-		)
-
-		if err != nil {
-			panic(err)
-		}
-		utils.VPrintf("Navigated to page %s", summary)
-
-		// Grab date/time of the event
-		var dateTimeStart time.Time
-		var dateTimeEnd time.Time
-		err = chromedp.Run(chromedpCtx,
-			chromedp.QueryAfter(".dtstart",
-				func(ctx context.Context, _ runtime.ExecutionContextID, nodes ...*cdp.Node) error {
-					if len(nodes) != 0 {
-						timeStamp, hasTime := nodes[0].Attribute("title")
-						if !hasTime {
-							return errors.New("event does not have a start time")
-						}
-						formattedTime, err := time.Parse(time.RFC3339, timeStamp)
-						if err != nil {
-							return err
-						}
-
-						dateTimeStart = formattedTime
-					}
-					return nil
-				}, chromedp.AtLeast(0),
-			),
-			chromedp.QueryAfter(".dtend",
-				func(ctx context.Context, _ runtime.ExecutionContextID, nodes ...*cdp.Node) error {
-					if len(nodes) != 0 {
-						timeStamp, hasTime := nodes[0].Attribute("title")
-						if !hasTime {
-							return errors.New("event does not have an end time")
-						}
-						formattedTime, err := time.Parse(time.RFC3339, timeStamp)
-						if err != nil {
-							return err
-						}
-
-						dateTimeEnd = formattedTime
-					}
-					return nil
-				}, chromedp.AtLeast(0),
-			),
-		)
-		if err != nil {
-			continue
-		}
-		utils.VPrintf("Scraped time: %s to %s ", dateTimeStart, dateTimeEnd)
-
-		// Grab Location of Event
-
-		// If p.location doesn't have children, then it's an virtual event
-		var location string = "Virtual Event"
-
-		err = chromedp.Run(chromedpCtx,
-			// Grab the name of the location
-			chromedp.QueryAfter("p.location > a",
-				func(ctx context.Context, _ runtime.ExecutionContextID, nodes ...*cdp.Node) error {
-					if len(nodes) != 0 {
-						// Location's name somehow contains leading space, so trim it
-						location = leadingSpaceRegex.ReplaceAllString(getNodeText(nodes[0]), "")
-					}
-					return nil
-				}, chromedp.AtLeast(0),
-			),
-			// Grab the address of the location (concatenated with the name)
-			chromedp.QueryAfter("p.location > span",
-				func(ctx context.Context, _ runtime.ExecutionContextID, nodes ...*cdp.Node) error {
-					if len(nodes) != 0 {
-						// There are cases where it doesn't show the address
-						if getNodeText(nodes[0]) != "" {
-							location += "\n" + getNodeText(nodes[0])
-						}
-					}
-					return nil
-				}, chromedp.AtLeast(0),
-			),
-		)
-		if err != nil {
-			continue
-		}
-		utils.VPrintf("Scraped location: %s, ", location)
-
-		// Get description of event
-		var description string = ""
-		err = chromedp.Run(chromedpCtx,
-			chromedp.QueryAfter(".em-about_description > p",
-				func(ctx context.Context, _ runtime.ExecutionContextID, nodes ...*cdp.Node) error {
-					// Concatenate all the sentences in the description together
-					for _, node := range nodes {
-						if getNodeText(node) != "" && getNodeText(node) != "\u00A0" {
-							description += getNodeText(node) + "\n\n"
-						}
-					}
-					return nil
-				}, chromedp.AtLeast(0),
-			),
-		)
-		if err != nil {
-			continue
-		}
-		utils.VPrintf("Scraped description: %s, ", description)
-
-		// Grab Event Type
-		var eventType []string = []string{}
-		err = chromedp.Run(chromedpCtx,
-			chromedp.QueryAfter(".filter-event_types > p > a",
-				func(ctx context.Context, _ runtime.ExecutionContextID, nodes ...*cdp.Node) error {
-					for _, node := range nodes {
-						eventType = append(eventType, getNodeText(node))
-					}
-					return nil
-				}, chromedp.AtLeast(0),
-			),
-		)
-		if err != nil {
-			panic(err)
-		}
-		utils.VPrintf("Scraped event type: %s", eventType)
-
-		// Grab Target Audience
-		targetAudience := []string{}
-		err = chromedp.Run(chromedpCtx,
-			chromedp.QueryAfter(".filter-event_target_audience > p > a",
-				func(ctx context.Context, _ runtime.ExecutionContextID, nodes ...*cdp.Node) error {
-					for _, node := range nodes {
-						targetAudience = append(targetAudience, getNodeText(node))
-					}
-					return nil
-				}, chromedp.AtLeast(0),
-			),
-		)
-		if err != nil {
-			panic(err)
-		}
-		utils.VPrintf("Scraped target audience: %s, ", targetAudience)
-
-		// Grab Topic
-		topic := []string{}
-		err = chromedp.Run(chromedpCtx,
-			chromedp.QueryAfter(".filter-event_topic > p > a",
-				func(ctx context.Context, _ runtime.ExecutionContextID, nodes ...*cdp.Node) error {
-					for _, node := range nodes {
-						topic = append(topic, getNodeText(node))
-					}
-					return nil
-				}, chromedp.AtLeast(0),
-			),
-		)
-		if err != nil {
-			panic(err)
-		}
-		utils.VPrintf("Scraped topic: %s, ", topic)
-
-		// Grab Event Tags
-		tags := []string{}
-		err = chromedp.Run(chromedpCtx,
-			chromedp.QueryAfter(".event-tags > p > a",
-				func(ctx context.Context, _ runtime.ExecutionContextID, nodes ...*cdp.Node) error {
-					for _, node := range nodes {
-						tags = append(tags, getNodeText(node))
-					}
-					return nil
-				}, chromedp.AtLeast(0),
-			),
-		)
-		if err != nil {
-			panic(err)
-		}
-		utils.VPrintf("Scraped tags: %s, ", tags)
-
-		// Grab Website
-		var eventWebsite string = ""
-		err = chromedp.Run(chromedpCtx,
-			chromedp.QueryAfter(".event-website > p > a",
-				func(ctx context.Context, _ runtime.ExecutionContextID, nodes ...*cdp.Node) error {
-					if len(nodes) != 0 {
-						href, hasHref := nodes[0].Attribute("href")
-						if !hasHref {
-							return errors.New("event does not have website")
-						}
-						eventWebsite = href
-					}
-					return nil
-				}, chromedp.AtLeast(0),
-			),
-		)
-		if err != nil {
-			continue
-		}
-		utils.VPrintf("Scraped website: %s, ", eventWebsite)
-
-		// Grab Department
-		var eventDepartment []string = []string{}
-		err = chromedp.Run(chromedpCtx,
-			chromedp.QueryAfter(".event-group > a",
-				func(ctx context.Context, _ runtime.ExecutionContextID, nodes ...*cdp.Node) error {
-					for _, node := range nodes {
-						eventDepartment = append(eventDepartment, getNodeText(node))
-					}
-					return nil
-				}, chromedp.AtLeast(0),
-			),
-		)
-		if err != nil {
-			panic(err)
-		}
-		utils.VPrintf("Scraped department: %s, ", eventDepartment)
-
-		// Grab Contact information
-		var contactInformationName string = ""
-		var contactInformationEmail string = ""
-		var contactInformationPhone string = ""
-		err = chromedp.Run(chromedpCtx,
-			chromedp.QueryAfter(".custom-field-contact_information_name",
-				func(ctx context.Context, _ runtime.ExecutionContextID, nodes ...*cdp.Node) error {
-					if len(nodes) != 0 {
-						contactInformationName = getNodeText(nodes[0])
-					}
-					return nil
-				}, chromedp.AtLeast(0),
-			),
-			chromedp.QueryAfter(".custom-field-contact_information_email > a",
-				func(ctx context.Context, _ runtime.ExecutionContextID, nodes ...*cdp.Node) error {
-					if len(nodes) != 0 {
-						emailHref, hasEmailHref := nodes[0].Attribute("href")
-						if !hasEmailHref {
-							return errors.New("event contact doesn't have email")
-						}
-						// Slicing the text to exclude "mailto:"
-						contactInformationEmail = emailHref[7:]
-					}
-					return nil
-				}, chromedp.AtLeast(0),
-			),
-			chromedp.QueryAfter(".custom-field-contact_information_phone",
-				func(ctx context.Context, _ runtime.ExecutionContextID, nodes ...*cdp.Node) error {
-					if len(nodes) != 0 {
-						contactInformationPhone = getNodeText(nodes[0])
-						if err != nil {
-							return err
-						}
-					}
-					return nil
-				}, chromedp.AtLeast(0),
-			),
-		)
-		if err != nil {
-			panic(err)
-		}
-		utils.VPrintf("Scraped contact name info: %s", contactInformationName)
-		utils.VPrintf("Scraped contact email info: %s", contactInformationEmail)
-		utils.VPrintf("Scraped contact phone info: %s", contactInformationPhone)
-
-		events = append(events, schema.Event{
-			Id:                 primitive.NewObjectID(),
-			Summary:            summary,
-			Location:           location,
-			StartTime:          dateTimeStart,
-			EndTime:            dateTimeEnd,
-			Description:        description,
-			EventType:          eventType,
-			TargetAudience:     targetAudience,
-			Topic:              topic,
-			EventTags:          tags,
-			EventWebsite:       eventWebsite,
-			Department:         eventDepartment,
-			ContactName:        contactInformationName,
-			ContactEmail:       contactInformationEmail,
-			ContactPhoneNumber: contactInformationPhone,
-		})
-	}
-
-	// Write event data to output file
-	fptr, err := os.Create(fmt.Sprintf("%s/events.json", outDir))
-	if err != nil {
-		panic(err)
-	}
-	encoder := json.NewEncoder(fptr)
-	encoder.SetIndent("", "\t")
-	encoder.Encode(events)
-	fptr.Close()
+	return parsedTime
 }
